### 2025-01-14 #7665 @editorialbot
[PRE REVIEW]: LongMemory.jl: Generating, Estimating, and Forecasting Long Memory Models in Julia
**Submitting author:** @everval (<a href="http://orcid.org/0000-0002-0337-8055">J. Eduardo Vera-Valdés</a>)
**Repository:** https://github.com/everval/LongMemory.jl
**Branch with paper.md** (empty if default branch): JOSS-submission
**Version:** v0.1.2
