### 2025-01-14 80d23a1
Fixed missing and incorrect DOIs
