+++ paper.bib
@@ -13,7 +13,7 @@
 
 @article{juliaprimer,
   author    = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  doi       = {https://doi.org/10.1137/141000671},
+  doi       = {10.1137/141000671},
   journal   = {SIAM review},
   number    = {1},
   pages     = {65--98},
@@ -54,6 +54,7 @@
     year = {2019},
     journal = {R package version 0.1.0},
     url = {https://CRAN.R-project.org/package=LongMemoryTS},
+    doi = {10.32614/CRAN.package.LongMemoryTS}
   }
 
 @article{VERAVALDES2024c,
