### 2025-01-16 #7665 @everval
Hi @lrnv thank you for the kind words.

Going through the list, I think that: Dom-Owens-UoB, SaranjeetKaur, sgeorge91, daviddewhurst, or ismailguzel could be great reviewers.

I already fixed the invalid DOI and I aim to retry adding the ones missing if they exist.
