### 2025-01-17 #7665 @lrnv
@iblislin @Dom-Owens-UoB, @SaranjeetKaur greetings :) Would any of you agree to review this submission for JOSS ? We cary on checklist-based reviews, you can read about our processes at https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
