### 2025-01-17 #7665 @lrnv
@PieterjanRobbe same thing: Would you be interested in reviewing this paper for JOSS ?
