### 2025-01-23 #7665 @lrnv
@gaelforget, @baxmittens, would any of you be interested in reviewing this paper for JOSS ?
