### 2025-01-23 #7665 @lrnv
@Santymax98, would you be interested in reviewing this paper for JOSS ?
