### 2025-01-23 #7665 @Santymax98
Hi dear Oskar, do you need a time limit? I have two weeks full of work… I
can start from de first of February

El El jue, 23 ene. 2025 a la(s) 9:01, Oskar Laverny <
***@***.***> escribió:

> @Santymax98 <https://github.com/Santymax98>, would you be interested in
> reviewing this paper for JOSS ?
>
