### 2025-01-23 #7665 @Santymax98
So yes, I can review it after February 1… without any problem 🚀

El El jue, 23 ene. 2025 a la(s) 9:48, Oskar Laverny <
***@***.***> escribió:

> Nope, there is no time limit @Santymax98 <https://github.com/Santymax98>,
> so you should be good to go.
>
