### 2025-01-23 #7665 @lrnv
@sgeorge91, @daviddewhurst, or @ismailguzel would any of you be willing to review this submission for JOSS ?
