### 2025-01-23 #7665 @lrnv
This would be enough
