### 2025-01-23 #7665 @ismailguzel
It sounds intriguing! I would be interested in being a reviewer for that submission to JOSS.
