### 2025-01-23 #7708 @editorialbot
[REVIEW]: LongMemory.jl: Generating, Estimating, and Forecasting Long Memory Models in Julia
**Submitting author:** @everval (<a href="http://orcid.org/0000-0002-0337-8055">J. Eduardo Vera-Valdés</a>)
**Repository:** https://github.com/everval/LongMemory.jl
**Branch with paper.md** (empty if default branch): JOSS-submission
**Version:** v1.0.0

@PieterjanRobbe & @Santymax98 & @ismailguzel & @baxmittens, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
