### 2025-01-23 #7708 @lrnv
@PieterjanRobbe, @Santymax98, @ismailguzel, @baxmittens greetings ! This is the review thread, in which we'll discuss and report everything happening around this review. I guess that all of you are familiar with JOSS reviews ? If not, please tell me and I'll give you more details. The first thing you have to do is to create a checklist for yourself by entering here `@editorialbot generate my checklist`.

We'll consider a 4-to-6 week window for the review, so for @baxmittens for example no problem, you can wait until mid-february to come here.
