### 2025-01-23 #7708 @lrnv
@everval have you ever published a JOSS paper ? The review works as follows: The four reviewers here are supposed to review your code, its documentation, the paper you wrote, and are encouraged to open issues and bugs onto your repository. Please take the time to consider what they say and answer them, the overall goal is of course to improve the quality of your proposal.
