### 2025-01-24 #7708 @everval
@lrnv This is my first ever submission to JOSS (first time, long time). And I am looking forward to the suggestions to improve the package and documentation.
