### 2025-02-05 #7708 @Santymax98
@everval Hi, could you explain to me why this happens?

```julia
using LongMemory
using YFinance, DataFrames, Dates, Statistics, StatsBase, StatsPlots
function get_prices_data(tickers::Vector{String}, start_date::String, end_date::String)
    # Descargar precios
    prices = get_prices.(tickers, startdt=start_date, enddt=end_date) |> (x -> DataFrame.(x)) |> (x -> vcat(x...))

    # Redondear fechas al día
    prices[!, :timestamp] = Dates.Date.(prices.timestamp)

    # Pivotear datos para dejar una columna por ticker
    prices_pivoted = unstack(prices, :timestamp, :ticker, :close)

    # Renombrar la columna de fechas
    rename!(prices_pivoted, :timestamp => :Date)

    return prices_pivoted
end

# Descarga de datos
tickers = ["^BVSP"]
data = get_prices_data(tickers, "1994-01-01", "2010-12-31")
rename!(data, Symbol("^BVSP") => :IBOVESPA)

function compute_absolute_returns(df::DataFrame)
    # Crear un nuevo DataFrame para los retornos absolutos
    returns = DataFrame(Date = df.Date[2:end])  # Excluir la primera fecha
    for col in names(df)[2:end]  # Iterar sobre las columnas excluyendo la columna Date
        # Calcular retornos absolutos
        returns[!, Symbol(string(col, "_ret_abs"))] = abs.(diff(df[!, col]))
    end
    # Filtrar filas donde haya NaN en las columnas numéricas (excluyendo la fecha)
    return dropmissing!(filter(row -> all(!isnan, row[2:end]), returns))
end
function compute_log_returns(df::DataFrame)
    # Crear DataFrame para los retornos logarítmicos
    returns = DataFrame(Date = df.Date[2:end])  # Excluir la primera fecha
    for col in names(df)[2:end]  # Iterar sobre las columnas excluyendo la columna Date
        # Calcular los retornos logarítmicos
        returns[!, Symbol(string(col, "_ret"))] = log.(df[2:end, col] ./ df[1:end-1, col])
    end
    return dropmissing!(filter(row -> all(!isnan, row[2:end]), returns))
end
# Calcular retornos absolutos para la serie IBOVESPA
data_returns = compute_absolute_returns(data)
log_returns = compute_log_returns(data)

acv_1 = autocovariance(data_returns[:,2], 30)
acv_2 = autocov(data_returns[:,2], 1:30)

acv_1 == acv_2
```
The result is `false`. It should be the same... on the other hand, isn't it better to omit these functions and make it dependent on StasBase? I think the functions are better optimized
