### 2025-02-13 #7708 @baxmittens
Hello all, hi @everval ,

looked over the Readme and the documentation and here are my first few comments:

1. I think the badges for the documentation and CI should be at the top of the readme at https://github.com/everval/LongMemory.jl

2. For me, most of the https://everval.github.io/LongMemory.jl/dev/ Home section is redundant, e.g. https://everval.github.io/LongMemory.jl/dev/#Generation
I think an outline or an index would be sufficient. I would recommend to look at https://documenter.juliadocs.org/stable/ as an example.

3. As far as I can see, community guidelines are missing. If not, can you point me to them @everval ?
4. I think the test routines in https://github.com/everval/LongMemory.jl/blob/master/test/runtests.jl are not sufficient. I think there should be actual testing of the core functionality of the package. I would expect to see some time series there, some estimation and some forecasting for benchmark examples.

While 1. and 2. are optional, 3. and 4. should be addressed at some point.

I also looked into the paper and from that some questions arose:

I cannot access Beran et al. Maybe @everval can point me to some other literature which can help me unterstand the theory?

In the first formula, there is a autocovariance function defined by an input lag. That is only possible if the stochastic process is
somewhat stationary. Is that a standard assumption of time series? I don't think so, or am I wrong?

I honestly do not understand the first equation:

γ_x(k) ≈ C_x k^(2*d-1) as k → ∞

What is d? And as far as I can see:
γ_x(k) = ∞ for d>1 if k → ∞ and γ_x(k) = 0 for d<0 if k → ∞

Am I missing something? Again, I think some further literature would be a good start for me to understand things.

Greetz Max
