### 2025-02-17 #7708 @PieterjanRobbe
Hi @everval, you seem to have an ArXiv paper [here](https://arxiv.org/pdf/2401.14077) that describes the functionality of your package. HI was wondering how this paper is different from the JOSS paper? Are you submitting this ArXiv paper to another journal? If so, what is its relation to the JOSS submission? Thanks!
