### 2025-02-17 everval/LongMemory.jl#13 @PieterjanRobbe
JOSS review
Hi @everval,

Thanks so much for a great package! The package appears to be well documented and has nice examples. It looks like it fills a niche in the Julia ecosystem. I particularly enjoyed reading the benchmark against `R` competitor packages, well done!

I will add my review comments as issues in this repository. You can close this issue if you want, or keep it open until all other issues have been addressed.

([breadcrumb](https://github.com/openjournals/joss-reviews/issues/7708))
