### 2025-02-17 everval/LongMemory.jl#14 @PieterjanRobbe
JOSS paper
These are my suggestions for the `JOSS` paper:
- I think it would be good to include a link to the code for generating Figure 1. Or, even better, add the code itself. Looking at your [example](https://everval.github.io/files/LM_notebook_illustration.html), I think you should be able to cover this with only a couple of lines.
- Page 2, line 26: there appears to be a spurious superscript "." after the `Julia` reference?
- Page 2, line 38: Perhaps you could add 1-2 sentences as to why this `Julia` package is better than the `R` packages you mention? Also, are there no `Python` packages for this?
- Page 3: You mention that you have accompanying "vignettes" for more examples of the package in use. Unless I'm mistaken, "vignette" is reserved for `R` notebooks, and it is not proper `Julia` terminology.

([breadcrumb](https://github.com/openjournals/joss-reviews/issues/7708#issuecomment-2656823714))
