### 2025-02-17 everval/LongMemory.jl#15 @PieterjanRobbe
Dependency on Plots
I noticed your package has a dependency on `Plots.jl`. This is a pretty heavy dependency to carry around. Perhaps it would make sense to use `PlotRecipes` instead, so you can remove the dependency on `Plots`?

([breadcrumb](https://github.com/openjournals/joss-reviews/issues/7708))
