### 2025-02-17 everval/LongMemory.jl#16 @PieterjanRobbe
Code coverage
I think the current test suite is not sufficient for JOSS standards. One think you could try to do is add coverage using [codecov](https://about.codecov.io/) or [coveralls](https://coveralls.io/) and try to get the coverage to an acceptable level, e.g., 90%?

([breadcrumb](https://github.com/openjournals/joss-reviews/issues/7708))
