+++ Project.toml
@@ -16,13 +16,13 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 
 [compat]
+Artifacts = "1"
 CSV = "0.10"
 DataFrames = "1"
-Artifacts = "1"
-Plots = "1"
 Distributions = "0.25"
 FFTW = "1"
 Optim = "1"
+Plots = "1"
 SpecialFunctions = "2"
 julia = "1.6"
 
