### 2025-02-25 #7708 @everval
Hi @baxmittens, thanks for your comments and suggestions.

I incorporate some of them already, while I keep working on the others.

- The badges should now appear top in the JOSS-submission branch. I plan to merge all the changes to main once all are done and approved.
- I added community guidelines at the end in Readme in the same branch. This is the first time that I add these, so please let me know if they are detailed enough. Happy to improve them further.
- Some literature freely available I can point to is the accepted author manuscript of one of my papers [here](https://vbn.aau.dk/en/publications/long-memory-fractional-integration-and-cross-sectional-aggregatio). An excellent short survey on the subject is *Long memory processes and fractional integration in econometrics* by Richard Baillie published in 1996. I found the latter can be accessed, along with other useful material, [here](http://long-memory.com/).
- The autocovariance can be defined regardless if the process is stationary or not. The main complication in the latter is that it may depend on time itself, so that no general conclusions can be made about it, besides probably that it goes to infinity or zero.
- Incidentally, that is  exactly what you correctly point out last. The parameter d is usualy called the long memory parameter. It can be shown that:
   - if d>1/2, the process is nonstationary and the autocovariance goes to infinity as the lag increases.
   - if |d|<1/2, the process is stationary and the autocovariance goes to zero. Nonetheless, the rate at which it goes to zero is hyperbolically, much slower than the usual geometric rate. The hyperbolic rate, as k increases, is characterised as k^a, while geometric decay will be a^k (with "a" between -1 and 1).
