### 2025-02-28 everval/LongMemory.jl#15 @everval
Hi @PieterjanRobbe thank you for the useful comments. I am actually intrigued about your suggestion to redude the package's dependencies. Nonetheless, I am not familiar with the alternative that you suggest. I tried googling PlotRecipes but could not find any useful information. Is there some documentation or example that you can point me to so that I can try and reduce the Plots dependency?
