### 2025-02-28 everval/LongMemory.jl#15 @PieterjanRobbe
Hi @everval, sorry I meant use [RecipesBase](https://github.com/JuliaPlots/Plots.jl/tree/v2/RecipesBase) instead of `Plots`.
