### 2025-03-03 07ce0e6
Added plot code to paper.md and corrected typo.
