+++ paper.md
@@ -36,9 +36,22 @@ As an example, the figure below shows the autocorrelation and periodogram (in lo
 
 As the figure shows, the autocorrelation function decays slowly and the periodogram diverges towards infinity near the origin. These are the features of long memory processes described in the definitions above. The `LongMemory.jl` package is concerned with methods for modeling data with this type of behavior.
 
+The following code generates the figure. The `autocorrelation_plot` and `periodogram_plot` functions are part of the `LongMemory.jl` package.
+
+```julia
+using LongMemory, Plots
+NileMin = NileData()
+theme(:ggplot2)
+p1 = plot( NileMin.Year , NileMin.NileMin, xlabel="Year" , ylabel = "Nile River minima" , legend = false )
+p2 = autocorrelation_plot(NileMin.NileMin, 50)
+p3 = periodogram_plot(NileMin.NileMin, slope = true)
+l = @layout [a; b c]
+plot(p1, p2, p3, layout = l, size = (700, 500) )
+```
+
 # Statement of need
 
-`LongMemory.jl` is a package for time series long memory modeling in Julia [@juliaprimer]˙. The package provides functions to generate long memory, estimate model parameters, and forecast. Generating methods include fractional differencing, stochastic error duration, and cross-sectional aggregation. Estimators include the classic ones used to estimate the Hurst effect, those inspired by log-periodogram regression, and parametric ones. Forecasting is provided for all parametric estimators. Moreover, the package adds plotting capabilities to illustrate long memory dynamics and forecasting. For some of the theoretical developments, `LongMemory.jl` provides the first publicly available implementation in any programming language. A notable feature of this package is that all functions are implemented in the same programming language, taking advantage of the ease of use and speed provided by Julia. Therefore, all code is accessible to the user. Multiple dispatch, a novel feature of the language, is used to speed computations and provide consistent calls to related methods. The package is related to the R [@R:lang] packages `LongMemoryTS` [@R:LongMemoryTS], `longmemo` [@R:longmemo], and `fracdiff` [@R:fracdiff].
+`LongMemory.jl` is a package for time series long memory modeling in Julia [@juliaprimer]. The package provides functions to generate long memory, estimate model parameters, and forecast. Generating methods include fractional differencing, stochastic error duration, and cross-sectional aggregation. Estimators include the classic ones used to estimate the Hurst effect, those inspired by log-periodogram regression, and parametric ones. Forecasting is provided for all parametric estimators. Moreover, the package adds plotting capabilities to illustrate long memory dynamics and forecasting. For some of the theoretical developments, `LongMemory.jl` provides the first publicly available implementation in any programming language. A notable feature of this package is that all functions are implemented in the same programming language, taking advantage of the ease of use and speed provided by Julia. Therefore, all code is accessible to the user. Multiple dispatch, a novel feature of the language, is used to speed computations and provide consistent calls to related methods. The package is related to the R [@R:lang] packages `LongMemoryTS` [@R:LongMemoryTS], `longmemo` [@R:longmemo], and `fracdiff` [@R:fracdiff].
 
 # Examples of Research Conducted with `LongMemory.jl`
 
