+++ paper.bib
@@ -48,6 +48,14 @@
   year    = {2020}
 }
 
+@article{BenchmarkToolsJL,
+  author    = {Jarrett Revels},
+  title     = {\pkg{BenchmarkTools.jl}},
+  url       = {https://juliaci.github.io/BenchmarkTools.jl/},
+  version   = {v1.4.0},
+  year      = 2023
+}
+
 @article{R:LongMemoryTS,
     title = {`LongMemoryTS`: Long Memory Time Series},
     author = {Leschinski, C.},
