+++ paper.md
@@ -30,7 +30,7 @@ where $f_x(\lambda)$ is the spectral density function and $C_f$ is a constant. A
 
 Both properties above can be analyzed graphically by plotting the autocorrelation and periodogram (an estimator of the spectral density), respectively.
 
-As an example, the figure below shows the autocorrelation and periodogram (in logs) for the Nile River minima data. The data are available in `LongMemory.jl` through `NileData`.
+As an example, the figure below shows the autocorrelation and periodogram (in logs) for the Nile River minima data. The data are available in `LongMemory.jl` through `NileData()`.
 
 ![Nile River minima (top), its autocorrelation function (bottom left), and log-periodogram (bottom right)](NileRiverMin.png)
 
@@ -42,7 +42,8 @@ The following code generates the figure. The `autocorrelation_plot` and `periodo
 using LongMemory, Plots
 NileMin = NileData()
 theme(:ggplot2)
-p1 = plot( NileMin.Year , NileMin.NileMin, xlabel="Year" , ylabel = "Nile River minima" , legend = false )
+p1 = plot( NileMin.Year , NileMin.NileMin, xlabel="Year", 
+      ylabel = "Nile River minima" , legend = false )
 p2 = autocorrelation_plot(NileMin.NileMin, 50)
 p3 = periodogram_plot(NileMin.NileMin, slope = true)
 l = @layout [a; b c]
@@ -51,7 +52,34 @@ plot(p1, p2, p3, layout = l, size = (700, 500) )
 
 # Statement of need
 
-`LongMemory.jl` is a package for time series long memory modeling in Julia [@juliaprimer]. The package provides functions to generate long memory, estimate model parameters, and forecast. Generating methods include fractional differencing, stochastic error duration, and cross-sectional aggregation. Estimators include the classic ones used to estimate the Hurst effect, those inspired by log-periodogram regression, and parametric ones. Forecasting is provided for all parametric estimators. Moreover, the package adds plotting capabilities to illustrate long memory dynamics and forecasting. For some of the theoretical developments, `LongMemory.jl` provides the first publicly available implementation in any programming language. A notable feature of this package is that all functions are implemented in the same programming language, taking advantage of the ease of use and speed provided by Julia. Therefore, all code is accessible to the user. Multiple dispatch, a novel feature of the language, is used to speed computations and provide consistent calls to related methods. The package is related to the R [@R:lang] packages `LongMemoryTS` [@R:LongMemoryTS], `longmemo` [@R:longmemo], and `fracdiff` [@R:fracdiff].
+`LongMemory.jl` is a package for time series long memory modeling in Julia [@juliaprimer]. The package provides functions to generate long memory, estimate model parameters, and forecast. Generating methods include fractional differencing, stochastic error duration, and cross-sectional aggregation. Estimators include the classic ones used to estimate the Hurst effect, those inspired by log-periodogram regression, and parametric ones. Forecasting is provided for all parametric estimators. Moreover, the package adds plotting capabilities to illustrate long memory dynamics and forecasting. For some of the theoretical developments, `LongMemory.jl` provides the first publicly available implementation in any programming language. A notable feature of this package is that all functions are implemented in the same programming language, taking advantage of the ease of use and speed provided by Julia. Therefore, all code is accessible to the user. Multiple dispatch, a novel feature of the language, is used to speed computations and provide consistent calls to related methods. 
+
+
+# Comparison to existing packages
+
[…]
