### 2025-03-03 everval/LongMemory.jl#14 @everval
Thanks for these. I made some improvements already:

- The code has been added after the plot. Great suggestion.
- Typo corrected.
- I added benchmarks against the most closely-related R packages. Interestingly, I cannot find a Python package for Long Memory. I added this note to the paper.
- I did not know about this context. Thanks for letting me know. I changed vignette to illustrative example.
