### 2025-03-03 ae1b0a7
Added harmonic processes, including inverse transform and Gregory coefficients
