### 2025-03-03 e999bbb
Additional tests for the harmonically weighted functions
