### 2025-03-07 0e78f56
Added no-detrending option to ELW
