### 2025-03-13 #7708 @baxmittens
@everval

> * The autocovariance can be defined regardless if the process is stationary or not. The main complication in the latter is that it may depend on time itself, so that no general conclusions can be made about it, besides probably that it goes to infinity or zero.

If you want to define your autocovariance by a lag $k=t_1-t_2$, i.e. $\gamma_x(k)$, rather than $\gamma_x(t_1,t_2)$, the process has to be at least weakly stationary. I also found that in the [link](https://vbn.aau.dk/en/publications/long-memory-fractional-integration-and-cross-sectional-aggregatio) you provided. Shouldn't that be reflected in the paper, or is it clear to all readers that we assume a stationary process if the autocovariance is defined by $k$?

> * Incidentally, that is  exactly what you correctly point out last. The parameter d is usualy called the long memory parameter. It can be shown that:
>
>   * if d>1/2, the process is nonstationary and the autocovariance goes to infinity as the lag increases.
>   * if |d|<1/2, the process is stationary and the autocovariance goes to zero. Nonetheless, the rate at which it goes to zero is hyperbolically, much slower than the usual geometric rate. The hyperbolic rate, as k increases, is characterised as k^a, while geometric decay will be a^k (with "a" between -1 and 1).

I think the domain/interval for $d$ should be given in the paper.
