### 2025-03-14 #7708 @PieterjanRobbe
Hi @lrnv, I've opened a few issues in the repo ([#13](https://github.com/everval/LongMemory.jl/issues/13), [#14](https://github.com/everval/LongMemory.jl/issues/14), [#15](https://github.com/everval/LongMemory.jl/issues/15) and [#16](https://github.com/everval/LongMemory.jl/issues/16)). I'm waiting for these issues to be addressed before continuing my review.
