### 2025-03-14 everval/LongMemory.jl#15 @baxmittens
I don't recommend going from `Plots ` to `RecipeBase`. `Plots` is superseded by `Makie`, which is also heavy in dependencies. At some point, there will be a lightweight interface for Makie. You could wait until then and then switch to `Makie` if you want to.
However, for the JOSS Paper, it is fine anyway, imo.
