### 2025-03-14 everval/LongMemory.jl#15 @PieterjanRobbe
Good point, closing for now.
