### 2025-03-19 everval/LongMemory.jl#14 @everval
Hi @PieterjanRobbe

Is it ok if I close this? I dont know if you have had the time to check the changes.
