### 2025-03-19 3118b0f
Specify stationary range
