+++ .github/workflows/CI.yml
@@ -20,7 +20,7 @@ jobs:
       matrix:
         version:
           - '1.6'
-          - '1.9'
+          - '1.10'
           - 'nightly'
         os:
           - ubuntu-latest
@@ -36,9 +36,12 @@ jobs:
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
       - uses: julia-actions/julia-processcoverage@v1
-      - uses: codecov/codecov-action@v3
+ #     - uses: codecov/codecov-action@v3
+ #       with:
+ #         files: lcov.info
+      - uses: codecov/codecov-action@v5
         with:
-          files: lcov.info
+          token: ${{ secrets.CODECOV_TOKEN }} 
   docs:
     name: Documentation
     runs-on: ubuntu-latest
