### 2025-03-19 c8554be
First batch of classic estimator tests
