### 2025-03-19 8bada41
typo in sstdk test
