+++ test/runtests.jl
@@ -1,16 +1,36 @@
 using LongMemory
-using Test, Plots
+using Test
 
 
 @testset "ClassicEstimators" begin
     @test smean(ones(10,1)) == 1
     @test autocovariance(ones(10,1),1) ≈ [0.0]
+    @test_throws ErrorException autocovariance(randn(10), 20)
     @test autocorrelation(collect(1:10),1) ≈ [1.0]
+    @test autocorrelation(collect(1:10),1; flag=true) ≈ [1.0]
+    @test_throws ErrorException autocorrelation(ones(10), 2)
     @test length(autocorrelation_plot(ones(10),1)) == 1
     @test length(log_variance_plot(randn(10,1); m = 4)) == 1
+    @test length(log_variance_plot(randn(10,1); m = 4, slope=true, slope2=true)) == 1
+    @test_throws ErrorException log_variance_plot(randn(2,1); m = 4)
     @test log_variance_est(collect(1:10); m = 4) ≈ 0.94596612
+    @test_throws ErrorException log_variance_est(collect(1:10); m = 20)
+    @test rescaled_range_est(collect(1:50)) ≈ -0.959937456
+    @test_throws ErrorException rescaled_range_est(collect(1:10))
+    @test length(rescaled_range_plot(collect(1:150))) == 1
+    @test_throws ErrorException rescaled_range_plot(collect(1:10))
+    @test rescaled_range(collect(1:30))[3] ≈ 33.0
+    @test_throws ErrorException rescaled_range(collect(1:10))
+    @test sstd(ones(10)) == 0.0
     @test sstdk(ones(10), 2) ≈ 0.0
+end
 
+@testset "DataExamples" begin
+    @test NileData()[1,1] == 622
+    @test NHTempData()[1,1] ≈ -0.901
+    @test length(NileDataPlot()) == 4
+    @test length(NHTempDataPlot()) == 4
+    @test length(LMPlot(ones(100))) == 4
 end
 
 @testset "LongMemory.jl" begin
@@ -30,8 +50,6 @@ end
     @test csa_var_vals(1, 1.4, 1.4) ≈ [2.1130846015858644]
     @test fi_cor_vals(1, 0.4) ≈ [1]
     @test csa_cor_vals(1, 1.4, 1.4) ≈ [1]
-    @test NileData()[1,1] == 622
-    @test NHTempData()[1,1] ≈ -0.901
     @test hitransform(hwfilter(ones(10))) ≈ ones(10)
     @test gregory_coeffs(3) ≈ [1/2; 1/12; 1/24]
     @test hitransform((zeros(10))) ≈ zeros(10)
