+++ test/runtests.jl
@@ -33,27 +33,60 @@ end
     @test length(LMPlot(ones(100))) == 4
 end
 
-@testset "LongMemory.jl" begin
-    @test fracdiff(ones(10,1),0) == ones(10,1)
-    @test fracdiff(ones(10,1),0.0) ≈ ones(10,1)
-    @test fracdiff(ones(3,1),0.5) ≈ [1; 0.5; 0.375]
-    @test csadiff(zeros(10,1),1.4,1.4) ≈ zeros(10,1)
-    @test arfima_gen(10,1,[],0,[];σ=0) ≈ zeros(10,1)
-    @test arfi_gen(10,0,[],0;σ=0) ≈ zeros(10,1)
-    @test arfi_gen(10,1,[],0;σ=0) ≈ ones(10,1)
-    @test fi_gen(10,0;σ=0) ≈ zeros(10,1)
+@testset "Forecasters" begin
     @test fi_ar_coefs(0.4,1) ≈ [0.4]
     @test fi_ar_coefs(0.2,1) ≈ [0.2]
     @test fi_ar_coefs(0.0,10) ≈ zeros(10,1)
+    @test gregory_coeffs(3) ≈ [1/2; 1/12; 1/24]
+    @test hitransform(hwfilter(ones(10))) ≈ ones(10)
+    @test hitransform((zeros(10))) ≈ zeros(10)
+    @test hitransform(ones(3)) ≈ [1; 1-1/2; 1-1/2-1/12]
+    @test fi_forecast(ones(10), 1, 0)[11] ≈ 0.0
+    @test fi_forecast(ones(10), 1, 0, 1)[11] ≈ 0.0
+    @test length(fi_forecast_plot(ones(10), 1, 0)) == 1
+    @test length(fi_forecast_plot(ones(10), 1, 0, 1)) == 1
+    @test csa_forecast(zeros(10), 2, 1.5, 1.5) == zeros(12,1)
+    @test csa_forecast(zeros(10), 2, 1.5, 1.5, 1)[1:12,1] == zeros(12)
+    @test length(csa_forecast_plot(zeros(10), 2, 1.5, 1.5)) == 1
+    @test length(csa_forecast_plot(zeros(10), 2, 1.5, 1.5, 1)) == 1
+    @test LongMemory.Forecasters.csa_ma_coefs(1.5, 1.5, 2) ≈ [1.0; sqrt(1/2)]
+    @test har_forecast(ones(100), 10)[101] ≈ 1.0
+    @test length(har_forecast_plot(ones(100), 10)) == 1
+    @test length(hwp_gen(10)) == 10
+end
+
+@testset "GeneratingFunctions" begin
+    @test fracdiff(ones(10,1),0) == ones(10,1)
+    @test fracdiff(ones(10,1),0.0) ≈ ones(10,1)
+    @test fracdiff(ones(3,1),0.5) ≈ [1; 0.5; 0.375]
+    @test fracdiff(ones(10), 1) == zeros(9)
+    @test csadiff(zeros(10,1),1.4,1.4) ≈ zeros(10,1)
+    @test length(csa_gen(10, 1.5, 1.5)) == 10
+    @test length(csa_gen(10, 2, 1.5, 1.5)) == 10
+    @test arfima_gen(10,1,[],0,[];σ=0) ≈ zeros(10,1)
+    @test length(arfima_gen(10,1,[1],0,[];σ=0)) == 10
+    @test length(arfima_gen(10,1,[1],0,[1];σ=0)) == 10
+    @test arfi_gen(10,0,[],0;σ=0) ≈ zeros(10,1)
+    @test arfi_gen(10,1,[],0;σ=0) ≈ ones(10,1)
[…]
