### 2025-03-20 06e4624
Up to testing LogPeriodEstimators
