### 2025-03-20 fc4226e
Getting the last 0.5%
