### 2025-03-20 everval/LongMemory.jl#16 @everval
Hi @PieterjanRobbe
I managed to add CodeCov and got it to 100%. It took me longer because it is the first time that I add code coverage, but I think it was well worth the effort. Let me know what you think.
