### 2025-03-20 #7708 @everval
@baxmittens You are right, that definition assumes weakly stationarity. So it is added to the paper.
Even though some of the tests included in the package are valid for the nonstationary region, a more general definition would be more cumbersome to write in the paper. So I will let it to the tests themselves.
Similar treatment has been done to the range of d.

Also, I added codecov and tests for all functions know.

Let me know what you think.
