### 2025-03-20 everval/LongMemory.jl#16 @PieterjanRobbe
Looks good to me!
