### 2025-03-25 #7708 @Santymax98
Hi @everval  @lrnv , sorry for the delay, I had finished. Great work, I hope in the future you can implement some additional things like volatility processes with ML. Good Luck!!!
