### 2025-03-25 #7708 @everval
Hi @baxmittens and @Santymax98. Thank you for reviewing the package.
And yes, @Santymax98, I plan to add FIGARCH-type processes to the package eventually. Nice to hear that it is something that would be useful.
