### 2025-03-27 1696aa8
Updated compat to last LTS version
