+++ Project.toml
@@ -24,7 +24,7 @@ FFTW = "1"
 Optim = "1"
 Plots = "1"
 SpecialFunctions = "2"
-julia = "1.6"
+julia = "1.10"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
