+++ v1.0.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"parametric/","page":"Parametric Estimation","title":"Parametric Estimation","text":"CurrentModule = LongMemory","category":"page"},{"location":"parametric/#Parametric-Estimators-for-Long-Memory","page":"Parametric Estimation","title":"Parametric Estimators for Long Memory","text":"","category":"section"},{"location":"parametric/","page":"Parametric Estimation","title":"Parametric Estimation","text":"Functions to estimate time series long memory models by parametric methods. Of particular interest is the ARFIMA model. Moreover, a method to estimate the HAR model, a specification usually used as a proxy for long memory dynamics, is also available.","category":"page"},{"location":"parametric/","page":"Parametric Estimation","title":"Parametric Estimation","text":"Modules = [ParametricEstimators]","category":"page"},{"location":"parametric/#LongMemory.ParametricEstimators","page":"Parametric Estimation","title":"LongMemory.ParametricEstimators","text":"ParametricEstimators\n\nThis module contains functions for estimating the parameters of the fractional differenced process à la ARFIMA and the CSA process.\n\nAuthor\n\nJ. Eduardo Vera-Valdés\n\n\n\n\n\n","category":"module"},{"location":"parametric/#LongMemory.ParametricEstimators.csa_cor_vals-Tuple{Int64, Real, Real}","page":"Parametric Estimation","title":"LongMemory.ParametricEstimators.csa_cor_vals","text":"csa_cor_vals(T::Int, p::Real, q::Real)\n\nComputes the autocorrelation function of the CSA process with parameters p and q at lags 0, 1, ..., T-1.\n\nArguments\n\nT::Int: The number of lags to compute.\np::Real: The first parameter of the CSA process.\nq::Real: The second parameter of the CSA process.\n\nOutput\n\nacf::Array: The autocorrelation function of the CSA process with parameters p and q at lags 0, 1, ..., T-1.\n\nNotes\n\nThis function uses csavarvals() to compute the autocovariance function and then normalizes by the variance (first computed value). \n\nExamples\n\njulia> csa_cor_vals(20, 0.4, 0.6)\n\n\n\n\n\n","category":"method"},{"location":"parametric/#LongMemory.ParametricEstimators.csa_llk-Tuple{Real, Real, Array}","page":"Parametric Estimation","title":"LongMemory.ParametricEstimators.csa_llk","text":"csa_llk(p::Real, q::Real, x::Array)\n\nComputes the log-likelihood of the CSA process with parameters p and q given the data x.\n\nArguments\n\np::Real: The first parameter of the CSA process.\nq::Real: The second parameter of the CSA process.\nx::Array: The data.\n\nOutput\n\nllk::Real: The log-likelihood of the CSA process with parameters p and q given the data x.\n\nNotes\n\nThis function computes the concentrated log-likelihood function of the CSA process with parameters p and q given the data x.\n\nExamples\n\njulia> csa_llk(1.4, 1.8, randn(100,1))\n\n\n\n\n\n","category":"method"},{"location":"parametric/#LongMemory.ParametricEstimators.csa_mle_est-Tuple{Array}","page":"Parametric Estimation","title":"LongMemory.ParametricEstimators.csa_mle_est","text":"csa_mle_est(x::Array)\n\nComputes the maximum likelihood estimate of the parameters p and q of the CSA process and the standard deviation of the CSA process given the data x.\n\nArguments\n\nx::Array: The data.\n\nOutput\n\np::Real: The maximum likelihood estimate of the first parameter of the CSA 
[…]
