+++ paper.bib
@@ -28,7 +28,8 @@
   journal = {R package version 1.5-2},
   title   = {`fracdiff`: Fractionally Differenced ARIMA Models},
   url     = {https://CRAN.R-project.org/package=fracdiff},
-  year    = {2022}
+  year    = {2022},
+  doi = {10.32614/CRAN.package.fracdiff},
 }
 
 @manual{R:lang,
@@ -45,7 +46,8 @@
   journal = {R package version 1.1-2},
   title   = {`longmemo`: Statistics for Long-Memory Processes and Related Functionality},
   url     = {https://CRAN.R-project.org/package=longmemo},
-  year    = {2020}
+  year    = {2020},
+  doi = {10.32614/CRAN.package.longmemo},
 }
 
 @article{BenchmarkToolsJL,
@@ -66,9 +68,12 @@
   }
 
 @article{VERAVALDES2024c,
-title = {Effects of the Paris Agreement and the COVID-19 Pandemic on Volatility Persistence of Stocks Associated with the Climate Crisis: A Multiverse Analysis},
-year = {2024},
-author = {Vera-Valdés, J.E., and Kvist, O.},
+  title = {Effects of the Paris Agreement and the COVID-19 Pandemic on Volatility Persistence of Stocks Associated with the Climate Crisis: A Multiverse Analysis},
+  year = {2025},
+  author = {Vera-Valdés, J.E., and Kvist, O.},
+  journal = {Advances in Econometrics},
+  volume = {Forthcoming},
+  url = {https://everval.github.io/publications/FinancialVolatilityAfterPA.html}
 }
 
 @article{vera-valdés2024,
