### 2025-03-31 #7708 @lrnv
@everval the title of the archive on zenodo is "everval/LongMemory.jl: v.1.0.0" while it should be the same as the paper's title, can you please fix that ?
