+++ paper.bib
@@ -28,6 +28,7 @@
   journal = {R package version 1.5-2},
   title   = {`fracdiff`: Fractionally Differenced ARIMA Models},
   url     = {https://CRAN.R-project.org/package=fracdiff},
+  doi     = {10.32614/CRAN.package.fracdiff},
   year    = {2022}
 }
 
@@ -40,20 +41,15 @@
   year         = {2023}
 }
 
-@article{R:longmemo,
-  author  = {Maechler, M. and Beran, J. and Whitcher, B.},
-  journal = {R package version 1.1-2},
-  title   = {`longmemo`: Statistics for Long-Memory Processes and Related Functionality},
-  url     = {https://CRAN.R-project.org/package=longmemo},
-  year    = {2020}
-}
-
-@article{BenchmarkToolsJL,
-  author    = {Jarrett Revels},
-  title     = {\pkg{BenchmarkTools.jl}},
-  url       = {https://juliaci.github.io/BenchmarkTools.jl/},
-  version   = {v1.4.0},
-  year      = 2023
+@ARTICLE{BenchmarkToolsJL,
+  author =	 {{Chen}, J. and {Revels}, J.},
+  title =	 "{Robust benchmarking in noisy environments}",
+  journal =	 {arXiv e-prints},
+  year =	 2016,
+  month =	 "Aug",
+  doi =	 {10.48550/arXiv.1608.04295},
+  archivePrefix ={arXiv},
+  url = {https://arxiv.org/abs/1608.04295}
 }
 
 @article{R:LongMemoryTS,
@@ -65,10 +61,13 @@
     doi = {10.32614/CRAN.package.LongMemoryTS}
   }
 
-@article{VERAVALDES2024c,
-title = {Effects of the Paris Agreement and the COVID-19 Pandemic on Volatility Persistence of Stocks Associated with the Climate Crisis: A Multiverse Analysis},
-year = {2024},
-author = {Vera-Valdés, J.E., and Kvist, O.},
+@article{VERAVALDES2025,
+  title = {Effects of the Paris Agreement and the COVID-19 Pandemic on Volatility Persistence of Stocks Associated with the Climate Crisis: A Multiverse Analysis},
+  year = {2025},
+  author = {Vera-Valdés, J.E., and Kvist, O.},
+  journal = {Advances in Econometrics},
+  volume = {Forthcoming},
+  url = {https://everval.github.io/publications/FinancialVolatilityAfterPA.html}
 }
 
 @article{vera-valdés2024,
