### 2025-03-31 #7708 @everval
@lrnv Oh sorry, perhaps I should not have run the check on DOIs? Just wanted to be sure I got them all. Sorry if I was not supposed to run it. 🙃
