### 2025-03-31 #7708 @lrnv
@everval No no of course the check is also there for you ! I was just telling you that i'll keep moving forward :). Btw i proposed a few minor typos correction and reformulations in the paper, can you take a look to decide on what you want to merge or not ?
