### 2025-03-31 9d7c629
Small edits, and value of d
