+++ paper.md
@@ -22,9 +22,9 @@ bibliography: paper.bib
 
 In time series analysis, the concept of *Long Memory* encompasses datasets including a strong dependency on past values. @Hurst1956 is one of the pioneering works on the subject from the field of Hydrology: while analyzing the flow of the Nile river, he noted that water reservoirs that do not account for its long-term dynamics were still at risk of overflowing. Long memory models are generally used in climate, finance, biology, economics, and many other fields. See @Beran2013 for a textbook on the subject.
 
-We say that a stationary time series $x_t$ has long memory with parameter $-1/2<d<1/2$ if it has autocovariance function $\gamma_x(k)$ that behaves as as:
+We say that a stationary time series $x_t$ has long memory with parameter $-1/2<d<1/2$ if it has an autocovariance function $\gamma_x(k)$ that behaves as:
 $$\gamma_x(k) \sim C_x k^{2d-1}\quad \textnormal{as}\quad k\to\infty, \label{def:cov}$$
-or if it has spectral density function $f_x(\lambda)$ that behaves as:
+or if it has an spectral density function $f_x(\lambda)$ that behaves as:
 $$f_x(\lambda)\sim C_f\lambda^{-2d}\quad \textnormal{as}\quad \lambda\to 0, \label{def:spectral}$$
 where both $C_x$ and $C_f$ are constants. Above equivalences $g(x)\sim h(x)$ as $x\to x_0$ holds when $g(x)/h(x)$ converges to $1$ as $x$ tends to $x_0$.
 
